refactor: extract dashboard layout to layouts/dashboard.py
